requirements.txt
fastapi
uvicorn
pydantic
pypdf
langchain
faiss-cpu
google-generativeai
sentence-transformers
python-multipart
langchain-community